Partially gutted IMU code to start writing to the DVL spec.
