+++ packages/drivers/dvl/Makefile
@@ -1,13 +1,13 @@
 CXX = g++
 CXXFLAGS = -std=c++11 -Wall -pedantic -O3
-HEADS = include/IMU.h include/types.h include/statics.h include/functions.h
-OBJS = bin/IMU.o bin/util.o
+HEADS = include/DVL.h include/types.h include/statics.h include/functions.h
+OBJS = bin/DVL.o bin/util.o
 
 .PHONY: all test clean
 
 .INTERMEDIATE: bin/util.o
 
-all: bin bin/IMU.o bin/config bin/baudrate
+all: bin bin/DVL.o bin/config bin/baudrate
 
 test: all bin/test
 	./bin/test /dev/ttyUSB0 115200
@@ -15,7 +15,7 @@ test: all bin/test
 bin:
 	mkdir bin
 
-bin/IMU.o: src/IMU.cpp $(HEADS)
+bin/DVL.o: src/DVL.cpp $(HEADS)
 	$(CXX) $(CXXFLAGS) -o $@ $< -c 
 
 bin/util.o: src/util.cpp include/util.h $(HEADS)
