+++ packages/drivers/dvl/include/util.h
@@ -1,16 +1,16 @@
-#ifndef IMU_UTIL_H
-#define IMU_UTIL_H
+#ifndef DVL_UTIL_H
+#define DVL_UTIL_H
 
-#include "IMU.h"
+#include "DVL.h"
 
 /** UNIX error code to return to the OS. */
-#define IMU_ERR 44
+#define DVL_ERR 45
 
 /** Prints the script usage arguments. */
 void printUsage();
 
 /** Converts a string into an IMUSpeed */
-IMU::IMUSpeed getBaudrate(const char*);
+DVL::DVLSpeed getBaudrate(const char*);
 
 /** Prints an exception and kills the program. */
 void printError(std::exception& e);
