+++ packages/drivers/dvl/src/util.cpp
@@ -12,26 +12,11 @@ void printUsage()
    exit(EX_USAGE);
 }
 
-IMU::IMUSpeed getBaudrate(const char* arg)
+DVL::DVLSpeed getBaudrate(const char* arg)
 {
+   throw DVLException("Unimplemented.");
    switch (atoi(arg))
    {
-      case 2400:
-         return IMU::k2400;
-      case 4800:
-         return IMU::k4800;
-      case 9600:
-         return IMU::k9600;
-      case 19200:
-         return IMU::k19200;
-      case 38400:
-         return IMU::k38400;
-      case 57600:
-         return IMU::k57600;
-      case 115200:
-         return IMU::k115200;
-      default:
-         return IMU::k0;
    }
 }
